// shared/calculators/interfaces/config_interfaces.ts
export interface TensorsToLandmarksConfig {
  numLandmarks: number;
  inputImageWidth: number;
  inputImageHeight: number;
  visibilityActivation: "none" | "sigmoid";
  flipHorizontally?: boolean;
  flipVertically?: boolean;
  normalizeZ?: number;
}